feat: introduce custom status identifiers
This PR introduces a new configuration field that allows the user to set their
own 'custom' status identifier other than
`continous-integration/bitrise`.
For me it is `app-resign/alpha-ipa` or
`app-resign/beta-ipa` to let my probot know when it needs to sync the GitHub Release
assets to the MDM provider for deployment.